Changed ExposedVideoData to use void* under Windows, fixes problems on 64bit systems. Some other 64bit warnings silenced. All changes provided by Kosuha.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@1313 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COBJMeshFileLoader.cpp
@@ -729,7 +729,7 @@ core::stringc COBJMeshFileLoader::copyLine(const c8* inBuf, const c8* bufEnd)
 			break;
 		++ptr;
 	}
-	return core::stringc(inBuf, ptr-inBuf+1);
+	return core::stringc(inBuf, (u32)(ptr-inBuf+1));
 }
